src/wingsight-stream_processor/requirements.txt
boto3==1.37.33
botocore==1.37.33
certifi==2025.1.31
charset-normalizer==3.4.1
dotenv==0.9.9
idna==3.10
inflect==7.5.0
jmespath==1.0.1
more-itertools==10.6.0
numpy==2.2.4
opencv-python==4.11.0.86
pika==1.3.2
pillow==11.2.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
s3transfer==0.11.4
six==1.17.0
sqlacodegen==3.0.0
SQLAlchemy==2.0.40
typeguard==4.4.2
typing_extensions==4.13.2
urllib3==2.4.0
